Solucionar Errores de tipos de datos y modificar la base de datos

+++ Parcial1-EmmanuelSuarez/Entidades/Productos.cs
@@ -13,14 +13,14 @@ namespace Parcial1_EmmanuelSuarez.Entidades
         public int ProductoId { get; set; }
         public string Descripcion { get; set; }
         public int Existencia { get; set; }
-        public float Costo { get; set; }
+        public decimal Costo { get; set; }
         
         public Productos()
         {
             ProductoId = 0;
             Descripcion = string.Empty;
             Existencia = 0;
-            Costo = 0.0f;
+            Costo = 0;
         }
     }
 }
